+++ src/costmap_inflator_imple.h
@@ -24,9 +24,11 @@ public:
     ~CostmapInflatorImple();
 
     sp<NcImageU8> inflate(NcImageU8* inputImg, Rect region);
+    void inflateInplace(NcImageU8* img, Rect region);
 
 private:
     void init();
+    void cleanup();
     CostValue computeCost(float distance);
 
     std::map<float, std::vector<CellData>> m_inflationCells;
@@ -36,7 +38,6 @@ private:
     int m_cachedInflationRadius;
 
     size_t m_cacheLineWidth;
-    const static size_t MAX_CACHE_SIZE = 64 * 64;
-    CostValue m_cachedCosts[MAX_CACHE_SIZE]; // m_cacheLineWidth * m_cacheLineWidth 2D array
-    float m_cachedDistances[MAX_CACHE_SIZE]; // m_cacheLineWidth * m_cacheLineWidth 2D array
+    CostValue* m_cachedCosts; // m_cacheLineWidth * m_cacheLineWidth 2D array
+    float* m_cachedDistances; // m_cacheLineWidth * m_cacheLineWidth 2D array
 };